Give all CNF data to GeneratePopCount
This adds a third element, `cnf`, to the JSON spec accepted by
`GeneratePopCount`. Provided by the frontend, this will facilitate
offloading the entire process of generating the CNF file _and_
executing `unigen` to the backend.

Values for `cnf` have this form:

```
"cnf": [
  [1, 2, -3],
  [3, 4, 1],
  ...
]
```

Where each inner list represents an "Or"ing of the variables.

Eventually, this JSON will be provided as input to a server on the
backend, which will generate the CNF file, run `unigen`, and then
return the solutions as the response.