Build docker image for server
This adds a `Dockerfile` to the project which builds a minimal docker
image capable of processing requests using `unigen`. The Dockerfile
itself contains instructions on how to build/push the image.

Images (only one so far) are available on docker hub [here][1].

I created a `sweetpea` organization on docker hub to manage the
images, when anyone else has a docker id, we can add them to the team.

[1]: https://hub.docker.com/r/sweetpea/server/tags/