Improve backend API
Backend requests will now require a `unigen` object of this form:

```json
{
  "unigen": {
    "support": <n>,
    "arguments": ["arg1", "arg2", ...]
  }
}
```

`support` is required, but `arguments` are optional. `support` should
be an integer, which indicates that variables from 1 to `n` constitute
an independent support for the formula, which will be included in the
generated DIMACS file.

`arguments` will be passed directly to the `unigen` process, which
will be useful as we determine how best to tune the `unigen`
sampling. (Time limited? Sample count limited?)

I'll have to build a new docker image for this, but I'm going to
change the frontend first so as to not cause undue (additional)
errors.