Add endpoint for solution count
I've felt the need for this a few times, just to give myself a better
idea of how variables and constraints impact the scale of the solution
space. Finally decided to implement it.

This will require adding the `sharpSAT` binary to the docker image,
which is low-overhead since `unigen` embeds the source in its own repo.
That will come in a subsequent commit once I've tested the new image.