Add endpoint for non-uniform solutions
While we work through issues with `unigen`, it would be really helpful
to our users if they could at least get _some_ solutions, even if
they're not uniform. That would at least allow them to start
experimenting with different designs, and find bugs sooner rather than
later.

This adds an endpoint to the server to facilitate this. Given a
request, and the number of desired solutions, the server will
repeatedly invoke `cryptominisat5` on the formula to build up a list
of unique solutions, with no guarantees about uniformity.

Uniqueness is guaranteed by adding the inverse of each solution to the
CNF formula.