Restructure to facilitate testing
I hit a hard-to-diagnose parsing bug involving whitespace (I was
forgetting to `strip` one of the intermediate strings) when parsing the
output from cryptominisat.

I decided to finally take some time to get a framework for unit testing
some of the server code to help me deal with these faster in the future.